requirements.txt
colorama==0.4.6
et_xmlfile==2.0.0
loguru==0.7.2
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
win32-setctime==1.1.0
